Merging r322081: ------------------------------------------------------------------------ r322081 | mgorny | 2018-01-09 06:44:04 -0800 (Tue, 09 Jan 2018) | 10 lines
[test] Fix tests to use more portable LLVM_ENABLE_ZLIB

The HAVE_LIBZ variable is not exported by LLVM, and therefore is not
available in stand-alone builds of other tools. Use LLVM_ENABLE_ZLIB
which is the name under which the effective value is exported.

Additional, use llvm_canonicalize_cmake_booleans() to make sure that
a correct (Python-safe) boolean value is passed down to lit.

Differential Revision: https://reviews.llvm.org/D41725
------------------------------------------------------------------------